+++ priv/templates/html/authorize.ex
@@ -1,95 +1,167 @@
 defmodule <%= base %>.Authorize do
+  @moduledoc """
+  Module to handle login and logout, and provide examples of user
+  authorization, with Openmaize.
+
+  This module provides the `handle_login` and `handle_logout` functions
+  to help with logging users in and out. The `handle_login` function
+  uses the user's role to redirect the user after logging in. If you are
+  not using roles, then you will need to edit this.
+
+  The rest of the documentation for this module will provide examples
+  of functions to authorize users that can be added to this module.
+
+  ## Examples of overriding the `action` function
+
+  One way of authorizing users is to customize the `action` function,
+  which is run before every route / function in the controller. Below
+  are examples of this kind of function. It's important to note that
+  all these functions will add another parameter to the functions in
+  your controller. For example, `def index(conn, params) do` will become
+  `def index(conn, params, user) do`.
+
+  A basic `action` function that just checks if the current_user is set.
+
+  In this file, add:
+
+      def auth_action(%Plug.Conn{assigns: %{current_user: nil}} = conn, _) do
+        unauthenticated conn
+      end
+      def auth_action(%Plug.Conn{assigns: %{current_user: current_user}} = conn, _) do
+        apply(module, action_name(conn), [conn, params, current_user])
+      end
+
+  In your controller file, import <%= base %>.Authorize and add:
+
+      def action(conn, _), do: auth_action conn, __MODULE__
+
+  A custom `action` function that checks the user id.
+
+  In this file, add:
+
+      def auth_action_id(%Plug.Conn{assigns: %{current_user: nil}} = conn, _) do
+        unauthenticated conn
+      end
+      def auth_action_id(%Plug.Conn{params: %{"user_id" => user_id} = params,
+        assigns: %{current_user: current_user}} = conn, module) do
+        if user_id == to_string(current_user.id) do
+          apply(module, action_name(conn), [conn, params, current_user])
+        else
+          unauthorized conn, current_user
+        end
+      end
+
+  In your controller file, import <%= base %>.Authorize and add:
+
+      def action(conn, _), do: auth_action conn, __MODULE__
+
+  A custom `action` function that checks the user role.
+
+  In this file, add:
+
+      def auth_action_role(%Plug.Conn{assigns: %{current_user: nil}} = conn, _, _) do
+        unauthenticated conn
+      end
+      def auth_action_role(%Plug.Conn{assigns: %{current_user: current_user},
+        params: params} = conn, roles, module) do
+        if current_user.role in roles do
+          apply(module, action_name(conn), [conn, params, current_user])
+        else
+          unauthorized conn, current_user
+        end
+      end
+
+  In your controller file, import <%= base %>.Authorize and add:
+
+      def action(conn, _), do: auth_action conn, ["admin", "user"], __MODULE__
+
+  ## Example plugs
+
+  These functions can be run before some or all of the functions in the
+  controller.
+
+  Plug that checks the user id.
+
+  In this file, add:
+
+      def id_check(%Plug.Conn{assigns: %{current_user: nil}} = conn, _opts) do
+        unauthenticated conn
+      end
+      def id_check(%Plug.Conn{params: %{"id" => id}, assigns: %{current_user:
+         %{id: current_id} = current_user}} = conn, _opts) do
+        id == to_string(current_id) and conn || unauthorized conn, current_user
+      end
+
+  In your controller file, import <%= base %>.Authorize and add:
+
+      plug :id_check, when action in [:show, :edit]
+
+  Plug that checks the user role.
+
+  In this file, add:
+
+      def role_check(%Plug.Conn{assigns: %{current_user: nil}} = conn, _opts) do
+        unauthenticated conn
+      end
+      def role_check(%Plug.Conn{assigns: %{current_user: current_user}} = conn, opts) do
+        roles = Keyword.get(opts, :roles, [])
+        current_user.role in roles and conn || unauthorized conn, current_user
+      end
+
+  In your controller file, import <%= base %>.Authorize and add:
+
+      plug :role_check ["admin"], when action in [:new, :create, :edit, :update]
+
+  In the example above, only users with the role `admin` can access
+  these resources.
+  """
 
   import Plug.Conn
   import Phoenix.Controller
-  #alias <%= base %>.{Repo, User}
 
   @redirects %{"admin" => "/admin", "user" => "/users", nil => "/"}
 
   @doc """
-  Custom action that can be used to override the `action` function in any
-  Phoenix controller.
+  Handle login using Openmaize.
 
-  This function checks for a `current_user` value. If there is no current_user,
-  the `unauthenticated` function is called.
+  If you are using two-factor authentication, uncomment the first commented
+  out section - the one that includes `render conn, "twofa.html", id: id`.
 
-  ## The current_user struct
-
-  This function produces a `current_user` with just the data from the JSON
-  Web Token. If you want the `current_user` to contain all the data about
-  the user from the database, uncomment the line `alias <%= base %>.{Repo, User}`
-  at the top of this module and the line `user = Repo.get(User, current_user.id)`
-  in this function.
-
-  ## Examples
-
-  First, import this module in the controller, and then add the following line:
-
-      def action(conn, _), do: authorize_action conn, __MODULE__
-
-  This command will only allow connections for users with the "admin" or "user"
-  role.
-
-  You will also need to change the other functions in the controller to accept
-  a third argument, which is the current user. For example, change:
-  `def index(conn, params) do` to: `def index(conn, params, user) do`
+  If you are using `remember me` functionality, uncomment the second section.
   """
-  def authorize_action(%Plug.Conn{assigns: %{current_user: nil}} = conn, _, _) do
-    unauthenticated conn
+  def handle_login(%Plug.Conn{private: %{openmaize_error: message}} = conn, _params) do
+    unauthenticated conn, message
   end
-  def authorize_action(%Plug.Conn{assigns: %{current_user: current_user},
-    params: params} = conn, module) do
-    #current_user = Repo.get(User, current_user.id)
-    apply(module, action_name(conn), [conn, params, current_user])
+  #def handle_login(%Plug.Conn{private: %{openmaize_otpdata: id}} = conn, _) do
+    #render conn, "twofa.html", id: id
+  #end
+  #def handle_login(%Plug.Conn{private: %{openmaize_user: %{id: id, role: role, remember: true}}} = conn,
+   #%{"user" => %{"remember_me" => "true"}}) do
+    #conn
+    #|> Openmaize.Remember.add_cookie(id)
+    #|> put_flash(:info, "You have been logged in")
+    #|> redirect(to: @redirects[role])
+  #end
+  def handle_login(%Plug.Conn{private: %{openmaize_user: %{id: id, role: role}}} = conn, _params) do
+    conn
+    |> put_session(:user_id, id)
+    |> put_flash(:info, "You have been logged in")
+    |> redirect(to: @redirects[role])
   end
 
   @doc """
-  Similar to `authorize_action`, but the user's role is also checked to
-  make sure it is in the list of authorized roles.
+  Handle logout using Openmaize.
 
-  This function checks for a `current_user` value, and if it finds it, it
-  then checks that the user's role is in the list of allowed roles. If
-  there is no current_user, the `unauthenticated` function is called, and
-  if the user's role is not in the list of allowed roles, the `unauthorized`
-  function is called.
-
-  ## The current_user struct
-
-  This function produces a `current_user` with just the data from the JSON
-  Web Token. If you want the `current_user` to contain all the data about
-  the user from the database, uncomment the line `alias <%= base %>.{Repo, User}`
-  at the top of this module and the line `user = Repo.get(User, current_user.id)`
-  in this function.
-
-  ## Examples
-
-  First, import this module in the controller, and then add the following line:
-
-      def action(conn, _), do: authorize_action_role conn, ["admin", "user"], __MODULE__
-
-  This command will only allow connections for users with the "admin" or "user"
-  role.
-
-  You will also need to change the other functions in the controller to accept
-  a third argument, which is the current user. For example, change:
-  `def index(conn, params) do` to: `def index(conn, params, user) do`
+  If you are using `remember me` functionality, uncomment the
+  `Openmaize.Remember.delete_rem_cookie` line.
   """
-  def authorize_action_role(%Plug.Conn{assigns: %{current_user: nil}} = conn, _, _) do
-    unauthenticated conn
-  end
-  def authorize_action_role(%Plug.Conn{assigns: %{current_user: current_user},
-    params: params} = conn, roles, module) do
-    if current_user.role in roles do
-      #current_user = Repo.get(User, current_user.id)
-      apply(module, action_name(conn), [conn, params, current_user])
-    else
-      unauthorized conn, current_user
-    end
+  def handle_logout(conn, _params) do
+    configure_session(conn, drop: true)
+    #|> Openmaize.Remember.delete_rem_cookie
+    |> put_flash(:info, "You have been logged out")
+    |> redirect(to: "/")
   end
 
-  @doc """
-  Redirect an unauthenticated user to the login page.
-  """
   def unauthenticated(conn, message \\ "You need to log in to view this page") do
     conn
     |> put_flash(:error, message)
@@ -97,109 +169,10 @@ defmodule <%= base %>.Authorize do
     |> halt
   end
 
-  @doc """
-  Redirect an unauthorized user to that user's role's page.
-
-  If you are not using roles, you will need to edit this function
-  before using it.
-
-  Each role has a redirect page associated with it, and these are set in the
-  `@redirects` module attribute in this file.
-  """
   def unauthorized(conn, current_user, message \\ "You are not authorized to view this page") do
     conn
     |> put_flash(:error, message)
     |> redirect(to: @redirects[current_user.role])
     |> halt
-  end
-
-  @doc """
-  Check, based on role, that the user is authorized to access this resource.
-
-  ## Examples
-
-  First, import this module, and then add the following line to the controller:
-
-      plug :role_check, [roles: "admin", "user"] when action in [:show, :edit]
-
-  This command will check the user's role for the `show` and `edit` routes.
-  """
-  def role_check(%Plug.Conn{assigns: %{current_user: nil}} = conn, _opts) do
-    unauthenticated conn
-  end
-  def role_check(%Plug.Conn{assigns: %{current_user: current_user}} = conn, opts) do
-    roles = Keyword.get(opts, :roles, [])
-    current_user.role in roles and conn || unauthorized conn, current_user
-  end
-
-  @doc """
-  Check, based on user id, that the user is authorized to access this resource.
-
-  ## Examples
-
-  First, import this module, and then add the following line to the controller:
-
-      plug :id_check when action in [:show, :edit, :update]
-
-  This command will check the user id for the `show`, `edit` and `update` routes.
-  """
-  def id_check(%Plug.Conn{assigns: %{current_user: nil}} = conn, _opts) do
-    unauthenticated conn
-  end
-  def id_check(%Plug.Conn{params: %{"id" => id}, assigns: %{current_user:
-     %{id: current_id} = current_user}} = conn, _opts) do
-    id == to_string(current_id) and conn || unauthorized conn, current_user
-  end
-
-  @doc """
-  Login and redirect to the user's role's page if successful.
-
-  If you are not using roles, you will need to edit this function
-  before using it.
-
-  ## Examples
-
-  Add the following line to the controller which handles login:
-
-      plug Openmaize.Login, [db_module: <%= base %>.OpenmaizeEcto] when action in [:login_user]
-
-  and then call `handle_login` from the `login_user` function:
-
-      def login_user(conn, params), do: handle_login(conn, params)
-
-  See the documentation for Openmaize.Login for all the login options.
-
-  ## Two factor authentication
-
-  This function can also be used for two factor authentication for any
-  user with `otp_required` set to true.
-  """
-  def handle_login(%Plug.Conn{private: %{openmaize_error: message}} = conn, _params) do
-    unauthenticated conn, message
-  end
-  def handle_login(%Plug.Conn{private: %{openmaize_otpdata:
-     {storage, uniq, id, _}}} = conn, _) do
-    render conn, "twofa.html", storage: storage, uniq: uniq, id: id
-  end
-  def handle_login(%Plug.Conn{private: %{openmaize_user: %{role: role}}} = conn, _params) do
-    conn
-    |> put_flash(:info, "You have been logged in")
-    |> redirect(to: @redirects[role])
-  end
-
-  @doc """
-  Logout and redirect to the home page.
-
-  ## Examples
-
-  Add the following line to the controller which handles logout:
-
-      plug Openmaize.Logout when action in [:logout]
-
-  and then call `handle_logout` from the `logout` function in the
-  controller.
-  """
-  def handle_logout(%Plug.Conn{private: %{openmaize_info: message}} = conn, _params) do
-    conn |> put_flash(:info, message) |> redirect(to: "/")
   end
 end